Finish template editor v1

+++ src/lib/database/requests/universes.requests.ts
@@ -1,6 +1,8 @@
 import {db} from "../index";
 import {universes} from "../schemas/universes";
 import {eq} from "drizzle-orm";
+import {deleteCharactersPanel} from "./characters.panels.requests";
+import {deleteMapsPanel} from "./maps.panels.requests";
 
 export async function getAllUniverses() {
     return db.select().from(universes);
@@ -23,6 +25,15 @@ export async function updateUniverse(id: number, {name, owners}: {name: string, 
 }
 
 export async function deleteUniverse(id: number) {
+    const universe = await getUniverseById(id)
+    if (universe.length < 1) {
+        return null
+    }
+
+    // delete all panels
+    await deleteCharactersPanel(universe[0].charactersPanel);
+    await deleteMapsPanel(universe[0].mapsPanel);
+
     return db.delete(universes).where(eq(universes.id, id))
 }
